refactor(sonar): extract shared vacuumAndGzip helper
Both db/snapshot.ts (takeSnapshot) and index/migrations/runner.ts
(writePreMigrationBackup) performed the identical VACUUM→chmod→gzip→
atomic-rename sequence. Extract it to db/vacuum-gzip.ts and replace
both call sites; remove the now-unused fs imports from runner.ts.
